backend/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Environment and configuration
python-dotenv==1.0.0

# Binance trading
python-binance==1.0.19

# Data validation
pydantic==2.5.0

# CORS
starlette==0.27.0

# Mathematical and scientific computing
numpy==1.24.3
pandas==2.0.3

# HTTP client for notifications (optional)
aiohttp==3.9.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Statistics and analysis
scipy==1.11.4

# Async utilities
asyncio-throttle==1.0.2
